refactor: `types` module
