+++ src/chains/optimism/index.ts
@@ -1,6 +1,6 @@
 import { optimism as optimismMetadata } from '@wagmi/chains';
-import { Chain } from '@/chains';
 import { sortedArrayByField } from '@/lib/utils';
+import { Chain } from '@/types';
 import { signatureTypes } from './signatureTypes';
 import { opcodes } from './vm/opcodes';
 import { precompiles } from './vm/precompiles';
