+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export type { Chain } from './chain';
+export type { Opcode } from './opcode';
+export type { Precompile, Predeploy } from './precompile';
+export type { SignatureType } from './signatureType';
